# Simple MSM perf improvements