--- a/src/msm-basic.ts
+++ b/src/msm-basic.ts
@@ -105,18 +105,21 @@ async function msmBasic(
 
   tic("accumulate and reduce");
   for (let { j, k, lstart, length } of chunksPerThread[thread]) {
-    tic(`chunk ${j}, length ${length} of partition ${k} - accumulate`);
+    tic();
     using _l = Field.local.atCurrentOffset;
     let buckets = Uint32Array.from(Field.local.getPointers(length, Curve.size));
     for (let l = 0; l < length; l++) setZero(buckets[l]);
 
     // accumulation
+    let nAdditions = 0;
+
     for (let i = 0, pi = pointPtr; i < N; i++, pi += size) {
       let l = pointsToBucket[k][i];
       let carry = l >>> 31;
       l &= 0x7f_ff_ff_ff;
       l -= lstart;
       if (l < 0 || l >= length) continue;
+      nAdditions++;
       let bucket = buckets[l];
       if (carry === 1) {
         subMixed(scratch, bucket, bucket, pi);
@@ -124,9 +127,15 @@ async function msmBasic(
         addMixed(scratch, bucket, bucket, pi);
       }
     }
-    toc();
+    let t = toc();
+    log(
+      `accumulate (chunk ${k}.${j}): ${t.toFixed(0)}ms, ${nAdditions} adds, ${(
+        (t / nAdditions) *
+        1e6
+      ).toFixed(1)}ns / add`
+    );
 
-    tic(`chunk ${j}, length ${length} of partition ${k} - reduce`);
+    tic(`reduce (chunk ${k}.${j})`);
     // reduce into sum
     let sum = chunkSumsPerPartition[k][j];
     reduceBucketsChunk(inputs, sum, buckets, lstart);