before:
```
> ./run scripts/run-msm-pallas.ts 16 16 --evaluate
msm (n=16)... 108ms ± 8ms
> ./run scripts/run-msm-ed-377.ts 16 16 --evaluate
msm (n=16)... 128ms ± 12ms
```

after:
```
> ./run scripts/run-msm-pallas.ts 16 16 --evaluate
msm (n=16)... 95ms ± 5ms
> ./run scripts/run-msm-ed-377.ts 16 16 --evaluate
msm (n=16)... 91ms ± 3ms
```